+++ src/main/java/com/alertops/team/service/TeamService.java
@@ -3,9 +3,9 @@ package com.alertops.team.service;
 import com.alertops.caching.Intent;
 import com.alertops.caching.IntentCache;
 import com.alertops.caching.IntentType;
-import com.alertops.global.GrantPermission;
-import com.alertops.global.Permission;
-import com.alertops.global.Role;
+import com.alertops.permissions.GrantPermission;
+import com.alertops.permissions.Permission;
+import com.alertops.permissions.Role;
 import com.alertops.security.AuthContext;
 import com.alertops.security.AuthContextHolder;
 import com.alertops.security.JwtUtil;
